+++ hw-3/simple-chat/src/index.js
@@ -1,5 +1,3 @@
 import './index.css';
 import './screen1_style.css';
 import './screen2_style.css';
-import './screen1.js';
-import './screen2.js';
