+++ hw-4/simple-chat/src/index.js
@@ -1,3 +1,6 @@
 import './index.css';
 import './screen1_style.css';
 import './screen2_style.css';
+import './main.js';
+import './screen1.js';
+import './screen2.js';
