+++ hw-4/simple-chat/src/main.js
@@ -1,0 +1,6 @@
+toChatList();
+
+let chatListButtons = document.getElementsByClassName("list-object");
+for (let button of chatListButtons) {
+    button.addEventListener("click", toChat);
+}
